Add offline support via a service worker
The service worker intercepts fetch requests so that when offline, the cached
version of site resources will be returned. This enables offline support in
manual testing with Chrome on Android and Firefox on Android, and Lighthouse
gives it a 92% overall score and confirms that it works offline.

Uses the boilerplate service worker registration code from
https://github.com/GoogleChrome/sw-precache/blob/master/demo/app/js/service-worker-registration.js
to load a service worker built on the sw-precache library, which handles
precaching of our smaller local static files, and dynamic runtime caching for
external resources and our large audio file.

The service worker generation gulp task runs after the styles and scripts
tasks.  If the scripts and styles don't exist when sw-precache runs, then they
won't get added to the precache list. Which makes it pretty hard to have this
operate offline.

Signed-off-by: Dan Scott <dan@coffeecode.net>